requirements.txt
faiss-cpu
numpy
langchain
langchain-google-genai
langchain-community
google-generativeai
pypdf